chore(mutation): updateUserDashboard renamed to updateEntity
