+++ src/containers/DetailsPanel/DetailsPanelHeader/DetailsPanelHeader.jsx
@@ -5,7 +5,7 @@ import copyToClipboard from '/src/helpers/copyToClipboard'
 import StackedThumbnails from '/src/pages/EditorPage/StackedThumbnails'
 import { classNames } from 'primereact/utils'
 import { isEqual, union } from 'lodash'
-import { useUpdateEntitiesMutation } from '/src/services/userDashboard/updateUserDashboard'
+import { useUpdateEntitiesMutation } from '/src/services/entity/updateEntity'
 import { toast } from 'react-toastify'
 import Actions from '/src/components/Actions/Actions'
 import FeedFilters from '../FeedFilters/FeedFilters'
