+++ src/pages/UserDashboardPage/UserDashboardTasks/UserDashboardList/UserDashboardList.jsx
@@ -4,7 +4,7 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { onCollapsedColumnsChanged, onTaskSelected } from '/src/features/dashboard'
 import { getFakeTasks, usePrefetchTask, useTaskClick } from '../../util'
-import { useUpdateEntitiesMutation } from '/src/services/userDashboard/updateUserDashboard'
+import { useUpdateEntitiesMutation } from '/src/services/entity/updateEntity'
 import { toast } from 'react-toastify'
 import getPreviousTagElement from '/src/helpers/getPreviousTagElement'
 import Shortcuts from '/src/containers/Shortcuts'
